shopify_recommender_ai-final/requirements.txt
pandas
numpy
psycopg2-binary
requests
Flask
python-dotenv